base configuration of peewee add user and fraction models
