+++ README.md
@@ -1,0 +1,6 @@
+ command to install requirements
+ 
+ `pip install --upgrade -r requirements.txt`
+ 
+ how to add autopep8 tool to pycharm
+ https://github.com/hscgavin/autopep8-on-pycharm
