+++ bot/models/application_record.py
@@ -1,0 +1,10 @@
+from peewee import *
+
+
+db = SqliteDatabase('replicant.db')
+db.connect()
+
+
+class BaseModel(Model):
+    class Meta:
+        database = db
