+++ bot/models/user.py
@@ -1,0 +1,13 @@
+from peewee import *
+from bot.models.application_record import BaseModel
+
+
+from bot.models.fraction import Fraction
+
+
+class User(BaseModel):
+    user_id = IntegerField(unique=True)
+    username = CharField(null=True)
+    first_name = CharField(null=True)
+    last_name = CharField(null=True)
+    fraction = ForeignKeyField(Fraction, backref='member')
