+++ requirements.txt
@@ -15,6 +15,6 @@ peewee==3.13.1
 pytz==2019.3
 PyYAML==5.1.2
 six==1.12.0
-SQLAlchemy==1.3.11
 tzlocal==2.0.0
 yarl==1.3.0
+autopep8
